pyproject.toml
[project]
name = "sheetbot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "GHCMaxim", email = "maxim080503@gmail.com" }
]
dependencies = [
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "google-api-python-client>=2.151.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sheetbot"]
